Added accel, gyro, hid to subfolders for organization, refactored to reflect path changes.
