+++ src/gyro/gyroCk.ts
