+++ src/hid/hidCk.ts
