+++ src/index.ts
@@ -1,7 +1,7 @@
 import DeferredPromise from "./DeferredPromise";
 import Chuck from "./Chuck";
-import HID from "./Hid";
-import Gyro from "./Gyro";
-import Accel from "./Accel";
+import HID from "./hid/Hid";
+import Gyro from "./gyro/Gyro";
+import Accel from "./accel/Accel";
 
 export { Chuck, HID, Gyro, Accel, DeferredPromise };
